refactor: useProductDetail.ts -> apis.ts 파일 이름 변경 및 위치 이동
